Cleaned up salpytools and ts_statemachine logic - namespace is now lsst.ts.linearStage - renamed ls module to hardware module - removed statemachine module - removed message and code returns as those are handled by salobj
