--- TODO
@@ -1,35 +1,0 @@
-Projects:
-  Documentation:
-    ☐ Add in ts_sal process
-    ☐ Add in link to pertinent parts of ts_sal user guide
-    ☐ Rewrite installation process
-  Intergrate state machine support:
-    ✔ Install salpytools @30m @done(18-07-13 08:49)
-    ☐ run salpytools using examples @30m
-    Learning:
-      The flow of salpytools:
-        ☐ ts_sal
-        ☐ salgenerator validate
-        ☐ salgenerator python
-    ☐ Initialize process from within library @1h
-    Create more modular setup:
-      ☐ Add in SalCommand class.
-   Add in XML support:
-    ✔ Figure out best way to create xml file @1h @done(18-07-12 14:50)
-    Add in support for commands:
-      ✔ Create xml for home command @15m @done(18-07-13 11:23)
-      ✔ Create xml for move_absolute command @15m @done(18-07-13 11:43)
-      ✔ Create xml for move_relative command @15m @done(18-07-13 12:03)
-      ✔ Create xml for get_position command @15m @done(18-07-13 12:22)
-    Add in support for events:
-      ✔ Figure out what is an event @30m @done(18-07-17 11:24)
-      ✔ create xml for home event @done(18-07-13 12:55)
-      ✔ create xml for move_absolute event @done(18-07-13 12:55)
-      ✔ create xml for move_relative event @done(18-07-13 13:02)
-      ✔ create xml for get_position event @1h @done(18-07-13 13:19)
-  
-Archive:
-  ✔ Get files committed to git. @10m @done(18-07-12 11:16) @project(Todo)
-  ✔ Get pip depedencies installed @1m @done(18-07-12 09:14) @project(Todo)
-  ✔ Create requirements.txt @30s @started(18-07-12 09:12) @done(18-07-12 09:13) @lasted(1m16s) @project(Todo)
-  ✔ Create setup.py @30m @started(18-07-12 11:16) @done(18-07-12 11:21) @lasted(5m52s) @project(Todo)
