+++ lsst/ts/linearStage/__init__.py
@@ -1,0 +1,1 @@
+# TODO: Add in package docstring
