--- python/lsst/ts/linearStage/__init__.py
@@ -1,3 +1,0 @@
-# TODO: Add in package docstring
-from .version import __version__
-__version__ = __version__
