+++ ups/ts_linearStage.table
@@ -13,5 +13,5 @@ setupRequired(pex_exceptions >= 4.6.0.0)
 # The following is boilerplate for all packages.
 # See Tech Note DMTN-001 for details on LSST_LIBRARY_PATH
 # Pure-Python packages only need the last line.
-envPrepend(PYTHONPATH, ${PRODUCT_DIR}/python)
+envPrepend(PYTHONPATH, ${PRODUCT_DIR}/lsst)
 envPrepend(PATH, ${PRODUCT_DIR}/scripts)
